+++ marimo/_messaging/types.py
@@ -1,10 +1,12 @@
 # Copyright 2024 Marimo. All rights reserved.
+from __future__ import annotations
+
 import abc
 import io
 from typing import Any, Dict, Optional, Tuple
 
-from marimo._ast.cell import CellId_t
 from marimo._messaging.mimetypes import KnownMimeType
+from marimo._types.ids import CellId_t
 
 # The message from the kernel is a tuple of message type
 # and a json representation of the message
