+++ marimo/_plugins/ui/_core/registry.py
@@ -13,9 +13,9 @@ else:
     from typing import TypeAlias
 
 from marimo._ast.app import _Namespace
-from marimo._ast.cell import CellId_t
 from marimo._plugins.ui._core.ui_element import UIElement
 from marimo._runtime.context import get_context
+from marimo._types.ids import CellId_t
 
 UIElementId = str
