+++ marimo/_runtime/app/common.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 
 from typing import Any, Dict, Tuple
 
-from marimo._ast.cell import CellId_t
+from marimo._types.ids import CellId_t
 
 OutputsType = Dict[CellId_t, Any]
 DefsType = Dict[str, Any]
