+++ marimo/_runtime/app/script_runner.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 import asyncio
 from typing import TYPE_CHECKING, Any, Callable, Iterator
 
-from marimo._ast.cell import CellId_t, CellImpl
+from marimo._ast.cell import CellImpl
 from marimo._dependencies.dependencies import DependencyManager
 from marimo._messaging.types import NoopStream
 from marimo._runtime.app.common import RunOutput
@@ -25,6 +25,7 @@ from marimo._runtime.patches import (
     create_main_module,
     patch_main_module_context,
 )
+from marimo._types.ids import CellId_t
 
 if TYPE_CHECKING:
     from marimo._ast.app import InternalApp
