+++ marimo/_runtime/runner/cell_runner.py
@@ -11,7 +11,7 @@ import traceback
 from dataclasses import dataclass
 from typing import TYPE_CHECKING, Any, Callable, Iterator, Optional, Union
 
-from marimo._ast.cell import CellId_t, CellImpl
+from marimo._ast.cell import CellImpl
 from marimo._ast.variables import unmangle_local
 from marimo._config.config import ExecutionType, OnCellChangeType
 from marimo._loggers import marimo_logger
@@ -34,6 +34,7 @@ from marimo._runtime.executor import (
     execute_cell_async,
 )
 from marimo._runtime.marimo_pdb import MarimoPdb
+from marimo._types.ids import CellId_t
 
 LOGGER = marimo_logger()
