+++ marimo/_runtime/scratch.py
@@ -1,6 +1,6 @@
 # Copyright 2024 Marimo. All rights reserved.
 from __future__ import annotations
 
-from marimo._ast.cell import CellId_t
+from marimo._types.ids import CellId_t
 
-SCRATCH_CELL_ID: CellId_t = "__scratch__"
+SCRATCH_CELL_ID = CellId_t("__scratch__")
