+++ marimo/_server/api/endpoints/datasources.py
@@ -9,9 +9,9 @@ from marimo import _loggers
 from marimo._runtime.requests import PreviewDatasetColumnRequest
 from marimo._server.api.deps import AppState
 from marimo._server.api.utils import parse_request
-from marimo._server.ids import ConsumerId
 from marimo._server.models.models import BaseResponse, SuccessResponse
 from marimo._server.router import APIRouter
+from marimo._types.ids import ConsumerId
 
 if TYPE_CHECKING:
     from starlette.requests import Request
