+++ marimo/_server/file_manager.py
@@ -9,7 +9,7 @@ from typing import Any, Dict, Optional
 from marimo import _loggers
 from marimo._ast import codegen
 from marimo._ast.app import App, InternalApp, _AppConfig
-from marimo._ast.cell import CellConfig, CellId_t
+from marimo._ast.cell import CellConfig
 from marimo._config.config import WidthType
 from marimo._runtime.layout.layout import (
     LayoutConfig,
@@ -22,6 +22,7 @@ from marimo._server.models.models import (
     SaveNotebookRequest,
 )
 from marimo._server.utils import canonicalize_filename
+from marimo._types.ids import CellId_t
 
 LOGGER = _loggers.marimo_logger()
