--- marimo/_server/ids.py
@@ -1,5 +1,0 @@
-# Copyright 2024 Marimo. All rights reserved.
-from __future__ import annotations
-
-SessionId = str
-ConsumerId = str
