+++ marimo/_server/session/session_view.py
@@ -5,7 +5,6 @@ import time
 from dataclasses import dataclass
 from typing import Any, Literal, Optional
 
-from marimo._ast.cell import CellId_t
 from marimo._data.models import DataSourceConnection, DataTable
 from marimo._messaging.cell_output import CellChannel, CellOutput
 from marimo._messaging.ops import (
@@ -27,6 +26,7 @@ from marimo._runtime.requests import (
     ExecutionRequest,
     SetUIElementValueRequest,
 )
+from marimo._types.ids import CellId_t
 from marimo._utils.lists import as_list
 from marimo._utils.parse_dataclass import parse_raw
