+++ marimo/_types/ids.py
@@ -1,0 +1,18 @@
+from __future__ import annotations
+
+from typing import NewType
+
+# TODO: enable NewType for better type checking
+# CellId_t = NewType("CellId_t", str)
+CellId_t = str
+
+# UIElementId = NewType("UIElementId", str)
+UIElementId = str
+
+# SessionId = NewType("SessionId", str)
+SessionId = str
+
+ConsumerId = NewType("ConsumerId", str)
+
+# VariableName = NewType("VariableName", str)
+VariableName = str
