+++ marimo/_utils/formatter.py
@@ -7,10 +7,10 @@ from typing import Dict
 
 from marimo import _loggers
 from marimo._dependencies.dependencies import DependencyManager
+from marimo._types.ids import CellId_t
 
 LOGGER = _loggers.marimo_logger()
 
-CellId_t = str
 
 CellCodes = Dict[CellId_t, str]
